+++ usage_test.go
@@ -2,6 +2,7 @@ package cmdline
 
 import (
 	"bytes"
+	"os"
 	"strings"
 	"testing"
 )
@@ -25,4 +26,26 @@ func Test_usage_output_with_extended_docs(t *testing.T) {
 		t.Error("incomplete")
 		t.Log(got)
 	}
+}
+
+func ExampleParser_usageHiddenPassword() {
+	cli := NewParser()
+	cli.args = []string{"adduser"}
+	_, opt := cli.Option("--uid").IntOpt(0)
+	opt.Doc("If not given, one is generated")
+	cli.Option("-p, --password",
+		"minimum 8 chars",
+		"hidden",
+	).String("secret")
+
+	cli.Usage().WriteTo(os.Stdout)
+	// output:
+	// Usage: adduser [OPTIONS]
+	//
+	// Options
+	//     --uid : 0
+	//         If not given, one is generated
+	//
+	//     -p, --password : "********"
+	//         minimum 8 chars
 }