+++ changelog.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [unreleased]
 
+- Hide default option values in usage by using a single docline with
+  the word "hidden"
 - Add Option.StringVar
 
 ## [0.14.0] 2023-07-14